+++ src/main/java/jarvis/command/EventCommand.java
@@ -1,8 +1,8 @@
 package jarvis.command;
 
-import jarvis.Storage;
+import jarvis.storage.Storage;
 import jarvis.task.TaskList;
-import jarvis.Ui;
+import jarvis.ui.Ui;
 import jarvis.exception.CommandParseException;
 import jarvis.task.EventTask;
