+++ src/main/java/jarvis/storage/Storage.java
@@ -1,4 +1,4 @@
-package jarvis;
+package jarvis.storage;
 
 import jarvis.exception.CommandParseException;
 import jarvis.exception.TaskIOException;
@@ -87,45 +87,5 @@ public class Storage {
         }
 
         return file;
-    }
-
-    /**
-     * Deserializes a string serial to a task.
-     *
-     * @param serial String serial from local storage,
-     * @return A task.
-     */
-    private Task deserialize(String serial) {
-        if (serial == null || serial.isBlank()) return null;
-
-        Scanner scanner = new Scanner(serial).useDelimiter("\\s*/\\s*");
-        String type = "";
-        boolean isDone = false;
-        String description = null;
-        String deadline = null;
-        String fromDateTime = null;
-        String toDateTime = null;
-
-        if (scanner.hasNext()) type = scanner.next();
-        if (scanner.hasNextBoolean()) isDone = scanner.nextBoolean();
-        if (scanner.hasNext()) description = scanner.next();
-
-        try {
-            switch (type) {
-                case "T":
-                    return new ToDoTask(description, isDone);
-                case "D":
-                    if (scanner.hasNext()) deadline = scanner.next();
-                    return new DeadlineTask(description, deadline, isDone);
-                case "E":
-                    if (scanner.hasNext()) fromDateTime = scanner.next();
-                    if (scanner.hasNext()) toDateTime = scanner.next();
-                    return new EventTask(description, fromDateTime, toDateTime, isDone);
-                default:
-                    return null;
-            }
-        } catch (CommandParseException e) {
-            return null;
-        }
     }
 }
